small fixes

+++ src/main/java/com/example/traineeship/mappers/TraineeshipPositionMapper.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.traineeship.domainmodel.TraineeshipPosition;
 
-public interface TraineeshipPositionMapper extends JpaRepository<TraineeshipPosition, Long>{
+public interface TraineeshipPositionMapper extends JpaRepository<TraineeshipPosition, Integer>{
 
 	Optional<TraineeshipPosition> findById(Integer positionId);
